requirements-5.txt
python-telegram-bot==21.3
flask==3.0.3
httpx==0.27.0
python-dotenv==1.0.1
